--- scrapers_library/data_portals/ckan/multi_portal_scraper/package.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class Package:
-    url: str = ""
-    title: str = ""
-    agency_name: str = ""
-    description: str = ""
